Add Keyboardio's production-tested kailh hotswap/solder hybrid footprint.
This footprint allows the assembler to hand-solder switches *or* use
Kailh hotswap sockets. The plated through-holes have most of the
through-hole annular ring on the top-side covered by solder mask to
reduce the chance of a bent switch stuck into a socket accidentally
working some of the time.

In addition, the plated through-holes should (we believe) provide some
additional strain relief on the pads when sockets are pressed hard.

The courtyard around the sockets is much tigher than on this library's
existing kailh hotswap footprints, allowing somewhat denser PCB designs.